* ✅  updates tests * fixes BackendSelect and SelectWithModal * Replaces TableModal with PaginatedTableModal
