+++ app/javascript/src/Common/index.js
@@ -4,7 +4,7 @@ export * from './components/CompactListCard'
 export * from './components/FancySelect'
 export * from './components/MicroPagination'
 export * from './components/NoMatchFound'
-export * from './components/PaginatedTableModal'
+export * from './components/TableModal'
 export * from './components/Select'
 export * from './components/SelectWithModal'
 export * from './components/Spinner'
